# controllers/init.js
//Constant declarations
const mongoose = require('mongoose');
const Story = require('../models/stories');

/**
 * Function which hardcodes some objects in mongo db and saves them
 */
exports.init = function () {

}